feat: use suspense component
